requirements.txt
# AI Core
google-generativeai>=0.8.0
openai>=1.12.0

# Media Processing
moviepy==1.0.3
ffmpeg-python>=0.2.0
Pillow>=10.2.0

# API Integrations
elevenlabs>=0.2.27
requests>=2.31.0

# Utils
python-dotenv>=1.0.0
pydantic>=2.6.0
tenacity>=8.2.3
